refactor(desk): create separate list pane content component
